Created a customerView
I created a customer view. it's supposed to be like the manager view except there is something present in the middle, there are options excluded, and some new ones that will be made. This one is not connected to a controller and also does not have any code written on it yet. Code must be written for it to be functional.