Got CreateReservationController.java working! (kinda) :)
This is kind of a massive edit to both CreateReservationController.java and Reservation.java . Made some changes to ManagerViewController and ManagerView.fxml (tried getting the Welcome text to show username as well). Created fx:id's for all buttons on customerview and fixed the weird title it had.

||Reservation.java Changes||
-Created Getters and Setters for FXML parameters in reservation.java
-ToDocument works (kinda)
-ToReservation works
-toString works

||CreateReservationController.java Changes||
-Cancel button works
-MakeReservation Button now takes all parameters and appends them to a document in MongoDB and jumps back to view

//CURRENT ISSUES\\
In mongodb in reservations, reservees is set to null. Couldn't figure out why it was being set to null even when setting App.currentUser. Also commented out the List reservatons in app.java due to the fact that it would force close the app cause reservees was set to null.