EmployeeView Changes and New Availability page (only for employeeview)
Made changes to Employeeview and added a table to show availability! However I need Jonas's help so friday it should be up and running!

Also made a new method in App.java that checks users permissions before switching scenes! so if you have a back button, you can just import app.java and call App.AccountTypeView();

Went through the files and implemented the new method.

That should be it for today, I just need to get that table displaying information!