EmployeeView Changes and Get Quote works!
Changed the resizable profile view since that was annoying. It was so annoying that we decided to just get rid of profile view for the employee and just replace it with the Hotel Hero Logo!

GET A QUOTE WORKS! (?)
it works but the logic for the amount of people still needs help. However, it do be done!

ISSUES:
Reservation view controller from employee view doesnt get the current hotel for some reason. so I gotta fix that